lifeline test works
